Add web routes for the application, including authentication, dashboard, user profile, master data, inventory, sales, purchasing, finance, and internal APIs

+++ database/migrations/2025_09_21_000008_create_invoice_items_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
             $table->id('item_id');
             $table->foreignId('invoice_id')->constrained('sales_invoices', 'invoice_id')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products', 'product_id');
-            $table->integer('quantity');
-            $table->integer('quantity_returned')->default(0);
+            $table->decimal('quantity', 15, 2);
+            $table->decimal('quantity_returned', 15, 2)->default(0);
             $table->decimal('price_per_unit', 15, 2);
             $table->decimal('hpp', 15, 2)->nullable()->default(0.00);
             $table->decimal('subtotal', 15, 2);
