+++ database/migrations/2025_09_23_000013_create_purchase_order_items_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id('item_id');
             $table->foreignId('po_id')->constrained('purchase_orders', 'po_id')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products', 'product_id');
-            $table->integer('quantity');
+            $table->decimal('quantity', 15, 2);
             $table->decimal('quantity_returned', 15, 2)->default(0);
             $table->decimal('price_per_unit', 15, 2);
             $table->decimal('subtotal', 15, 2);
